% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContactsInternal.R
\name{getContactsInternal}
\alias{getContactsInternal}
\title{A function rather aimed at developers}
\usage{
getContactsInternal(x, y, dt)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
