% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContactsPatient.R
\name{getContactsPatient}
\alias{getContactsPatient}
\title{A function rather aimed at developers}
\usage{
getContactsPatient(x, y, dt)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
