% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatientContactsToDF.R
\name{PatientContactsToDF}
\alias{PatientContactsToDF}
\title{A function rather aimed at developers}
\usage{
PatientContactsToDF(a)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
