% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisitContactTrace.R
\name{VisitContactTrace}
\alias{VisitContactTrace}
\title{A contact tracing application built to consume healthcare encounter data.}
\usage{
VisitContactTrace(viewer = "browser", browserURL = NULL)
}
\arguments{
\item{viewer}{By default is opens up application in Browser. Options ('browser','pane','dialog')}

\item{browserURL}{By default is NULL and is used to provide custom browser path for shiny app.}
}
\description{
A contact tracing application built to consume healthcare encounter data.
}
